+++ srcs/offset.c
@@ -6,21 +6,21 @@
 /*   By: kemethen <kemethen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/07 23:04:30 by kemethen          #+#    #+#             */
-/*   Updated: 2019/02/08 13:18:09 by kemethen         ###   ########.fr       */
+/*   Updated: 2019/02/12 16:16:09 by kemethen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fillit.h"
 
-short		offset_x(t_tetri *t)
+int		offset_x(t_tetri *t)
 {
-	short	*tab;
-	short	tmp;
+	int		*tab;
+	int		tmp;
 	int		i;
 
 	tmp = 0;
 	i = 0;
-	if (!(tab = (short *)malloc(sizeof(short) * 4)))
+	if (!(tab = (int *)malloc(sizeof(int) * 4)))
 		return (-1);
 	tab[0] = t->a.x;
 	tab[1] = t->b.x;
@@ -41,15 +41,15 @@ short		offset_x(t_tetri *t)
 	return (tmp);
 }
 
-short		offset_y(t_tetri *t)
+int		offset_y(t_tetri *t)
 {
-	short	*tab;
-	short	tmp;
+	int		*tab;
+	int		tmp;
 	int		i;
 
 	tmp = 0;
 	i = 0;
-	if (!(tab = (short *)malloc(sizeof(short) * 4)))
+	if (!(tab = (int *)malloc(sizeof(int) * 4)))
 		return (-1);
 	tab[0] = t->a.y;
 	tab[1] = t->b.y;
